% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexecution.R
\name{asDense}
\alias{asDense}
\title{Converts a sparse matrix or vector to a dense sparse matrix or vector}
\usage{
asDense(x)
}
\description{
Converts a sparse matrix or vector to a dense sparse matrix or vector
}
