% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storageConversion.R
\name{asSparse}
\alias{asSparse}
\title{Converts a dense matrix or vector to a sparse matrix or vector}
\usage{
asSparse(x)
}
\description{
Converts a dense matrix or vector to a sparse matrix or vector
}
