% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_nCompiler.R
\name{cpp_nCompiler}
\alias{cpp_nCompiler}
\title{Call Rcpp's C++ compilation system for nCompiler-generated content}
\usage{
cpp_nCompiler(
  Rcpp_packet,
  dir = file.path(tempdir(), "nCompiler_generatedCode"),
  cacheDir = file.path(tempdir(), "nCompiler_RcppCache"),
  env = parent.frame(),
  write = TRUE,
  compile = TRUE,
  ...
)
}
\arguments{
\item{dir}{Directory for writing C++ code files.}

\item{cacheDir}{Directory to use for Rcpp's cache.}

\item{env}{Environment for loading results of compilation.}

\item{...}{Any additional argument provided will be passed to
\code{sourceCpp_nCompiler}.}

\item{cpp_nCompilerPacket}{An object containing content for .h and .cpp
files.}
}
\value{
If one function is being compiled, \code{cpp_nCompiler} returns
    an R function that calls the compiled C++ internally is
    returned.  If multiple functions are being compiled,
    \code{cpp_nCompiler} returns a list of such functions.  If no
    functions are being compiled, \code{cpp_nCompiler} returns
    \code{NULL}.
}
\description{
Call Rcpp's C++ compilation system for nCompiler-generated content
}
\details{
\code{cpp_nCompiler} is not intended to be called directly.
    Instead functions such as \code{nCompile} (TBD),
    \code{nCompile_nFunction}, and \code{nCompile_nClass}
    should be used.
}
