% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_DLL.R
\name{createLocalDLLpackage}
\alias{createLocalDLLpackage}
\title{Create (but don't install) files for nCompLocal package}
\usage{
createLocalDLLpackage(dir = ".", force = TRUE)
}
\arguments{
\item{dir}{Directory in which nCompLocal package files will be generated.}
}
\description{
Write files and compile static library for nCompLocal package, whose purpose is to 
hold a static library (DLL) that on-the-fly compilation can link to.
}
\details{
This function is called internally from \cite{\link{buildLocalDLLpackage}}.  
See there for explanation of the local DLL.
}
\seealso{
\cite{\link{buildLocalDLLpackage}}, \cite{\link{installLocalDLLpackage}}
}
