% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_nOption}
\alias{get_nOption}
\title{Get nCompiler Option}
\usage{
get_nOption(x)
}
\arguments{
\item{x}{a character string holding an option name}
}
\value{
The value of the option.
}
\description{
Allow the user to get the value of a global _option_
that affects the way in which nCompiler operates
}
