% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_DLL.R
\name{installLocalDLLpackage}
\alias{installLocalDLLpackage}
\title{Install nCompLocal package}
\usage{
installLocalDLLpackage(lib, source.dir = ".")
}
\arguments{
\item{lib}{location to install package.  This defaults to the first element of `.libPaths()`.  `lib` matches behavior, 
and is passed to, the \code{lib} argument of \link{install.packages}.}

\item{source.dir}{directory where nCompLocal source is located.}
}
\description{
Install the nCompLocal package, whose purpose is to hold a static library for linking from on-the-fly compilations.
}
\details{
This function is used internally by This function is called internally from \cite{\link{buildLocalDLLpackage}}.
See there for explanation of the local DLL.
}
\seealso{
\cite{\link{buildLocalDLLpackage}}, \cite{\link{createLocalDLLpackage}}
}
