% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Utils.R
\name{isCompiledNCgenerator}
\alias{isCompiledNCgenerator}
\title{Determine if an object is a compiled nClass generator}
\usage{
isCompiledNCgenerator(x)
}
\arguments{
\item{x}{Object to be inspected}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Determine if an object is a compiled nClass generator, returned by a call to \link{nCompile_nClass} with \code{interface} set to \code{"full"} or \code{"both"}.
}
