% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Utils.R
\name{isNC}
\alias{isNC}
\title{Determine if an object is a nClass object}
\usage{
isNC(x)
}
\arguments{
\item{x}{Object to be inspected.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Determine if an object is a nClass object, returned by a nClass generator.
}
