% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Utils.R
\name{isNCgenerator}
\alias{isNCgenerator}
\title{Determine if an object is a nClass generator}
\usage{
isNCgenerator(x)
}
\arguments{
\item{x}{Object to be inspected}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Determine if an object is a nClass generator, returned by a call to \code{\link{nClass}}.
}
