% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NF_Utils.R
\name{isNF}
\alias{isNF}
\title{Check if an object is a nFunction}
\usage{
isNF(x)
}
\arguments{
\item{x}{Object to be checked.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if an object is a nFunction
}
