% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_nCompiler.R
\name{listRcppPackets}
\alias{listRcppPackets}
\title{List available saved RcppPackets in a directory}
\usage{
listRcppPackets(dir)
}
\arguments{
\item{dir}{Directory to search for saved packets}
}
\value{
Character vector of packet names found
}
\description{
List available saved RcppPackets in a directory
}
