% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_nCompiler.R
\name{loadRcppPacket}
\alias{loadRcppPacket}
\title{Load an RcppPacket from disk files}
\usage{
loadRcppPacket(dir, name)
}
\arguments{
\item{dir}{Directory containing the packet files}

\item{name}{Base name of the packet files to load}
}
\value{
An RcppPacket object (list)
}
\description{
Load an RcppPacket from disk files
}
