% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexecution.R
\name{nBacksolve}
\alias{nBacksolve}
\title{Upper-triangular solver}
\usage{
nBacksolve(U, b)
}
\arguments{
\item{U}{an upper-triangular matrix}

\item{b}{a matrix or vector}
}
\description{
In a \code{nFunction}, \code{nBacksolve} is identical to 
\code{backsolve} that solves the upper-triangular system U %*% x = b when
b represents a matrix or vector
}
\details{
This function is similar to R's \code{\link{backsolve}} function, 
  but can be used in a nFunction and compiled using \code{nCompile}.
}
