% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexecution.R
\name{nChol}
\alias{nChol}
\title{Compute the cholesky decomposition of a matrix}
\usage{
nChol(x)
}
\arguments{
\item{x}{a symmetric matrix}
}
\description{
In a \code{nFunction}, \code{nChol} is identical to \code{chol}
}
\details{
This function is similar to R's \code{\link{diag}} function, but 
  can be used in a nFunction and compiled using \code{nCompile}.
}
