% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Compile.R
\name{nCompile_nClass}
\alias{nCompile_nClass}
\title{Compile a nClass.}
\usage{
nCompile_nClass(
  NC,
  dir = file.path(tempdir(), "nCompiler_generatedCode"),
  cacheDir = file.path(tempdir(), "nCompiler_RcppCache"),
  env = parent.frame(),
  compileInfo = NULL,
  control = list(),
  interface = c("full", "generic", "both"),
  project_env = new.env(),
  ...
)
}
\arguments{
\item{NC}{A nClass generator (returned from a call to \link{nClass}).}

\item{dir}{Directory where generated C++ will be written.}

\item{cacheDir}{Directory to be used for Rcpp cache.}

\item{env}{Environment to be used for loading results of compilation.}

\item{control}{List of control settings for compilation.  See...}
}
\value{
Generator of objects of the compiled version of class
    \code{NC}.  These will use C++ objects internally for compiled
    data and methods.
}
\description{
Generates C++ for the compilable data and methods of a nClass,
manages C++ compilation of them and returns a generator for obejcts
of the compiled class.
}
