% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NF_Compile.R
\name{nCompile_nFunction}
\alias{nCompile_nFunction}
\title{Compile a nFunction}
\usage{
nCompile_nFunction(
  NF,
  dir = file.path(tempdir(), "nCompiler_generatedCode"),
  cacheDir = file.path(tempdir(), "nCompiler_RcppCache"),
  env = parent.frame(),
  compileInfo = NULL,
  control = list(),
  project_env = new.env(),
  ...
)
}
\arguments{
\item{NF}{A nFunction to be compiled.}

\item{dir}{Directory where generated C++ will be written.}

\item{cacheDir}{Directory to be used for Rcpp cache.}

\item{env}{Environment for loading compiled results.}

\item{control}{= List of control parameters.}
}
\value{
A function that will internally call the compiled C++ code.
}
\description{
Compiler a nFunction via automatic generation and compilation of C++.
}
