% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexecution.R
\name{nDiag}
\alias{nDiag}
\title{Extract or replace the diagonal of matrix}
\usage{
nDiag(x)
}
\arguments{
\item{x}{a numeric or complex matrix}
}
\description{
In a \code{nFunction}, \code{nDiag} is identical to \code{diag}
}
\details{
This function is similar to R's \code{\link{diag}} function, but 
  can be used in a nFunction and compiled using \code{nCompile}.
}
