% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_utils.R
\name{nDim}
\alias{nDim}
\alias{dim}
\title{return sizes of an object whether it is a vector, matrix or array}
\usage{
nDim(obj)
}
\arguments{
\item{obj}{objects for which the sizes are requested}
}
\value{
a vector of sizes in each dimension
}
\description{
R's regular \code{dim} function returns NULL for a vector.  It is useful to have this function that treats a vector similarly to a matrix or array.  Works in R and nCompiler.  In nCompiler \code{dim} is identical to \code{nDim}, not to R's \code{dim}
}
\examples{
x <- rnorm(4)
dim(x)
nDim(x)
y <- matrix(x, nrow = 2)
dim(y)
nDim(y)

}
\author{
nCompiler development team
}
