% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NF_Utils.R
\name{NFinternals}
\alias{NFinternals}
\alias{`NFinternals<-`}
\alias{NFinternals<-}
\title{Get or set internals of a nFunction (advanced use only).}
\usage{
NFinternals(NF)

NFinternals(NF) <- value
}
\arguments{
\item{NF}{A \code{nFunction}.}

\item{value}{Object of class \code{NF_InternalsClass}.}
}
\description{
Get or set internals of a nFunction (advanced use only).
}
