% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexecution.R
\name{nForwardsolve}
\alias{nForwardsolve}
\title{Lower-triangular solver}
\usage{
nForwardsolve(L, b)
}
\arguments{
\item{L}{a lower-triangular matrix}

\item{b}{a matrix or vector}
}
\description{
In a \code{nFunction}, \code{nForwardsolve} is identical to 
\code{forwardsolve} that solves the lower-triangular system L %*% x = b when
b represents a matrix or vector
}
\details{
This function is similar to R's \code{\link{forwardsolve}} function, 
  but can be used in a nFunction and compiled using \code{nCompile}.
}
