% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexecution.R
\name{nLogdet}
\alias{nLogdet}
\title{Compute the log-determinant of a matrix}
\usage{
nLogdet(x)
}
\arguments{
\item{x}{a square matrix}
}
\description{
In a \code{nFunction}, \code{nLogdet} is identical to \code{logdet}
}
\details{
This function is similar to R's \code{\link{diag}} function, but 
  can be used in a nFunction and compiled using \code{nCompile}.
}
