% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexecution.R
\name{nRep}
\alias{nRep}
\title{Replicate Elements of Vectors and Lists}
\usage{
nRep(x, ...)
}
\arguments{
\item{x}{a vector (of any mode including a list) or a factor or 
(for rep only) a POSIXct or POSIXlt or Date object; or an S4 object 
containing such an object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
In a \code{nFunction}, \code{nRep} is identical to \code{base::rep}
}
\details{
This function is similar to R's \code{\link{rep}} function, but 
  can be used in a nFunction and compiled using \code{nCompile}.
}
