% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_NC_Predefined.R
\name{OptimControlList}
\alias{OptimControlList}
\title{nClass containing control list options for stats::optim function}
\description{
Defines C++ bindings for the optim function's control list
}
\details{
Create a new `OptimControlList` object.

Set members to default values
}
\examples{
control <- OptimControlList$new()
control$initToDefaults()
}
