% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_DLL.R
\name{requireLocalDLLpackage}
\alias{requireLocalDLLpackage}
\title{Find and load, or possibly create, nCompLocal package.}
\usage{
requireLocalDLLpackage(lib, buildIfMissing = TRUE)
}
\arguments{
\item{lib}{optional location where nCompLocal is installed}

\item{buildIfMissing}{if \code{TRUE}, this function attempt to create nCompLocal if it can't be found.}
}
\description{
nCompLocal is a package created and installed on each user's machine, sometimes or each R
session.  It contains a static library for nCompiler to link to.
}
\details{
This function is called internally every time nCompiler is ready to call the C++ compiler for
generated code.
}
\seealso{
\link{setup_nCompLocal} for how to create and install nCompLocal once for an
R installation, so that nCompLocal can be re-used than than re-created in every R session.
}
