% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_nCompiler.R
\name{saveRcppPacket}
\alias{saveRcppPacket}
\title{Save an RcppPacket to disk as a set of files}
\usage{
saveRcppPacket(RcppPacket, dir, name = NULL)
}
\arguments{
\item{RcppPacket}{The RcppPacket object to save}

\item{dir}{Directory to save the packet files in}

\item{name}{Base name for the packet files (defaults to filebase from packet)}
}
\value{
Invisibly returns the directory path where files were saved
}
\description{
Save an RcppPacket to disk as a set of files
}
