% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_nOption}
\alias{set_nOption}
\title{Set nCompiler Option}
\usage{
set_nOption(x, value, listName = NULL)
}
\arguments{
\item{x}{a character string holding an option name}

\item{value}{the new value of the option}

\item{listName}{an optional character string with the name of a list in which to find \code{x}}
}
\value{
\code{value}
}
\description{
Allow the user to set the value of a global _option_
that affects the way in which nCompiler operates
}
