% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NF_derivs.R
\name{setup_wrt}
\alias{setup_wrt}
\title{Computes the value, Jacobian, and Hessian of a given  
\code{nFunction} method.}
\usage{
setup_wrt(nFxn = NA, dropArgs = NA, wrt = NULL, NC = NULL)
}
\arguments{
\item{nFxn}{a call to a compiled or uncompiled \code{nFunction} method with
arguments included.}

\item{dropArgs}{a vector of integers specifying any arguments to \code{nFxn}
that derivatives should not be taken with respect to.  For example,
\code{dropArgs = 2} means that the second argument to \code{nFxn} will not
have derivatives taken with respect to it.  Defaults to an empty vector.}

\item{wrt}{a character vector of either: names of function arguments to take
derivatives with respect to.  If left empty, derivatives will be taken
with respect to all arguments to \code{nFxn}.}

\item{NC}{An nClass generator (returned from a call to \code{\link{nClass}})
which has a method corresponding to the one found in \code{nFxn}. Only
required when \code{nFxn} is a compiled method call from a generic or full
interface.}
}
\value{
an integer vector of the positions in the flattened input of the
  elements with respect to which to differentiate \code{nFxn}.
}
\description{
Computes the value, Jacobian, and Hessian of a given  
\code{nFunction} method.
}
