% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_utils.R
\name{writeCode}
\alias{writeCode}
\title{Write unlisted code generated from nCompiler cpp definitions.}
\usage{
writeCode(x, ...)
}
\arguments{
\item{x}{A (potentially nested) list of C++ code in character
strings.}

\item{...}{Argument provided in \code{...} will be passed to
\code{writeLines}}
}
\description{
This is not intended to be called directly but is useful for
debugging nCompiler's C++ output.
}
