% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexecution.R
\name{asSparse}
\alias{asSparse}
\title{Converts a dense matrix or vector to a sparse matrix or vector}
\usage{
asSparse(x, prune = TRUE)
}
\arguments{
\item{x}{object to convert to sparse representation}

\item{prune}{TRUE to remove 0's from an object if it is already stored in a 
sparse format}
}
\description{
Converts a dense matrix or vector to a sparse matrix or vector
}
