% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{buildPackage}
\alias{buildPackage}
\title{Build and install packages written by writePackage}
\usage{
buildPackage(package.name, dir = ".", lib, load = TRUE, roxygenize = FALSE)
}
\arguments{
\item{package.name}{Character string. The name of the package to be built,
corresponding to the argument of the same name in writePackages.}

\item{dir}{Character string. Path to the parent directory containing the main
package directory. By default, the current working directory is used.}

\item{lib}{Character string, optional. Path to the directory where the
package will be installed. See the lib.loc argument in
`install.packages()`.}

\item{load}{Logical, default TRUE. Should the package be attached once
installed?}
}
\description{
Build and install packages written by writePackage
}
