% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{documentNClass}
\alias{documentNClass}
\title{Produce a roxygen2 documentation text block for an nClass}
\usage{
documentNClass(
  obj = NULL,
  name,
  title,
  description = NULL,
  fields = list(),
  CMethodsDescriptions = list(),
  CMethodsParams = list(),
  otherRoxygen = NULL,
  processWhitespace = TRUE,
  checkAgainstObj = TRUE
)
}
\arguments{
\item{obj}{The nClass generator being documented. Only required if
checkAgainstObj is TRUE, ignored otherwise.}

\item{name}{The nClass name. Used in the `@name` field of roxygen2
documentation.}

\item{title}{a descriptive title for the nClass. Used in the `@title` field
of roxygen2 documentation}

\item{description}{A text block giving a full and clear description of the
nClass. For use in the `@description` field of roxyygen2 documentation. If
processWhitespacpe is TRUE, any whitespace in the description will be
ignored and the code block will be reformatted for style.}

\item{fields}{A named list of documentation for C-ready public fields
belonging to the nClass. Fields are values stored internally by the class
and that are publicly accessible with `obj$field`. List names should
exactly match field names, while the elements of the list should be
character strings describing the field.}

\item{CMethodsDescriptions}{A named list of descriptions for the nClass's
methods. List names should exactly match method names, while the elements
of the list should be character strings describing the method.}

\item{CMethodsParams}{A named list of named lists. The names in
CMethodsParams should match exactly the names of the methods of the nClass.
Each element of CMethodsParams should itself be a named list, with names
matching parameters of the appropriate function, and elements being text
strings describing those parameters.}

\item{processWhitespace}{Logical, default TRUE. Can whitespace be modified to
maintain style?}

\item{checkAgainstObj}{Logical, default TRUE. Should the input be compared
against the actual nClass generator to confirm that all elements have been
documented appropriately?}
}
\description{
Builds out a text block in roxygen2 format given the inputs. The
  user provides the name, title, and description, as well as a named list of
  descriptions of fields and a named list of methods descriptions.
  Optionally, the function can process whitespace and perform a check against
  the nClass generator object, reporting whether all fields and methods were
  described.
}
\seealso{
\link[roxygen2]{roxygenize}
}
