% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{erasePackage}
\alias{erasePackage}
\title{Erase packages generated by `nCompiler`.}
\usage{
erasePackage(package.name, dir, nCompilerOnly = TRUE, uninstall = FALSE)
}
\arguments{
\item{package.name}{The name of the package to be erased}

\item{dir}{The directory containing the package to be erased}

\item{nCompilerOnly}{Logical, default TRUE. If TRUE, extra checks are 
performed to make sure the target directory is a package generated by
nCompiler. If FALSE, checks are still performed to ensure the target
looks like a package.}

\item{unintall}{Logical, default FALSE. If TRUE, the package is uninstalled
as well as erased.}
}
\description{
Erases all files associated with a package written out by the
  `writePackage()` function in nCompiler.
}
