% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_nParse.R
\name{nDeparse}
\alias{nDeparse}
\title{Convert a nCompiler parse tree to an R parse tree or text.}
\usage{
nDeparse(code, indent = "", toR = FALSE)
}
\arguments{
\item{code}{An \code{exprClass} object, representing a.nCompiler parse
tree.}

\item{indent}{Amount of indentation in text output if \code{toR} is
\code{FALSE}}

\item{toR}{If \code{TRUE}, the returned object will be an R parse
tree (similar to what is returned from \code{quote()} or
\code{call()}.  If \code{FALSE}, the returned object will be
text.}
}
\value{
Either an R parse tree or text, depending on the value of \code{toR}.
}
\description{
Convert a nCompiler parse tree, specifically an \code{exprClass}
object, to an R parse tree or text.
}
