% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NF_derivs.R
\name{nDerivs}
\alias{nDerivs}
\title{nFunction Derivatives}
\usage{
nDerivs(nFxn = NA, order = c(0, 1, 2), dropArgs = NA, wrt = NULL,
  NC = NULL)
}
\arguments{
\item{nFxn}{a call to a compiled or uncompiled \code{nFunction} method with
arguments included.}

\item{order}{an integer vector with values within the set {0, 1, 2},
corresponding to whether the function value, Jacobian, and Hessian should
be returned respectively.  Defaults to \code{c(0, 1, 2)}.}

\item{dropArgs}{a vector of integers specifying any arguments to \code{nFxn}
that derivatives should not be taken with respect to.  For example,
\code{dropArgs = 2} means that the second argument to \code{nFxn} will not
have derivatives taken with respect to it.  Defaults to an empty vector
and is ignored if \code{wrt} is provided.}

\item{wrt}{a vector of either: names of function arguments to take
derivatives with respect to or integers for the positions in the flattened
input vector of the elements with respect to which to differentiate.  If
left empty, derivatives will be taken with respect to all arguments to
\code{nFxn}.}

\item{NC}{An nClass generator (returned from a call to \link{nClass})
which has a method corresponding to the one found in \code{nFxn}.  Only
required when \code{nFxn} is a compiled method call from a generic or full
interface.}
}
\value{
a \code{list} with elements \code{value}, \code{gradient}, and
  \code{hessian}.
}
\description{
Computes the value, Jacobian, and Hessian of a given  
\code{nFunction} or compiled \code{nClass} method.
}
\details{
Derivatives for uncompiled nFunctions are calculated using the
  \code{numDeriv} package.  If this package is not installed, an error will
  be issued.  Derivatives for matrix valued arguments will be returned in
  column-major order.
}
