% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rexecution.R
\name{nEigen}
\alias{nEigen}
\title{Spectral Decomposition of a Matrix}
\usage{
nEigen(x)
}
\arguments{
\item{x}{a numeric or complex matrix whose spectral decomposition is to be 
computed. Logical matrices are coerced to numeric.}
}
\description{
In a \code{nFunction}, \code{nEigen} is identical to \code{eigen}
}
\details{
This function is similar to R's \code{\link{eigen}} function, but 
  can be used in a nFunction and compiled using \code{nCompile}.
}
