% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_nParse.R
\name{nParse}
\alias{nParse}
\title{Create.nCompiler parse tree from R code.}
\usage{
nParse(code, caller = NULL, callerArgID = numeric(), recursing = FALSE)
}
\arguments{
\item{code}{R code object such as returned by \code{quote()} or
\code{call()}.}

\item{caller}{exprClass object in which \code{code} is an argument.
Used only recursively.}

\item{callerArgID}{Integer index of which argument \code{code} is
in its \code{caller}.  Used only recursively.}

\item{recursing}{\code{TRUE} only if called by recursion.}
}
\value{
An \code{exprClass} object.
}
\description{
Create nCompiler parse tree, an annotated bidirectional syntax tree
used by the.nCompiler compiler, from R code.
}
