% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Serialize.R
\name{read_nClass}
\alias{read_nClass}
\title{Read an an nClass object saved by \code{save_nClass}}
\usage{
read_nClass(file, package.name, dir = tempdir(), lib = .libPaths()[1])
}
\arguments{
\item{file}{The (probably .rds) file to which the nClass object was written}

\item{package.name}{The name of the package to which the nClass}

\item{dir}{The directory in which the package was written (probably not going
to be a part of the final product)}

\item{lib}{The lib folder where the package will be quietly installed
(probably not going to be a part of the final product)}
}
\description{
Read an an nClass object saved by \code{save_nClass}
}
