% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Serialize.R
\name{save_nClass}
\alias{save_nClass}
\title{Save an instance of an nClass object across sessions}
\usage{
save_nClass(
  ncObj,
  ncDef,
  file,
  package.name,
  dir = tempdir(),
  lib = .libPaths()[1]
)
}
\arguments{
\item{ncObj}{A compiled instance of an nClass object. Right now this only
works for}

\item{ncDef}{The uncompiled nClass generator used to define ncObj's nClass}

\item{file}{The filename to which the serialized object will be written as
RDS, probably with extension ".rds"}

\item{package.name}{The name of the package that'll be generated to store the
nClass definition (probably not going to be a part of the final product)}

\item{dir}{The directory in which the package will be written (probably not
going to be a part of the final product)}

\item{lib}{The lib folder where the package will be quietly installed
(probably not going to be a part of the final product)}
}
\description{
TBW
}
