% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_nOption}
\alias{set_nOption}
\title{Set nCompiler Option}
\usage{
set_nOption(name, value)
}
\arguments{
\item{value}{the new value of the option}

\item{x}{a character string holding an option name}
}
\value{
\code{value}
}
\description{
Allow the user to set the value of a global _option_
that affects the way in which nCompiler operates
}
