% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packaging.R
\name{writePackage}
\alias{writePackage}
\title{Create packages containing compiled elements}
\usage{
writePackage(
  ...,
  package.name,
  dir = ".",
  control = list(),
  modify = get_nOption("modifyPackageContents"),
  roxygen = list(),
  nClass_full_interface = TRUE
)
}
\arguments{
\item{...}{One or more nClass constructor and nFunction objects to be
compiled into the package.}

\item{package.name}{Character string. The name of the package to be written.}

\item{dir}{Character string. Path to the parent directory in which the main
package directory will be created. If not provided, default is the current
working directory.}

\item{control}{A named list of control options or a named list of named lists
of control parameters for each object to be compiled. See Details for more.}

\item{modify}{Logical, default FALSE. Should writing proceed even if the
package exists? If TRUE, elements provided for compilation are added to the
existing package. Existing package elements with the same name as new
elements are overwritten, while elements without name conflicts are
retained. To change the default across multiple calls, use
`set_nOption("modifyPackageContents" = TRUE)`}

\item{roxygen}{A list of roxygen entries corresponding to the objects being
compiled, indicated either by matching names or (if all objects are
documented) by order. The elements of the list are either character strings
or lists produced by documentNClass.}

\item{nClass_full_interface}{}

\item{roxygenize}{Logical, default FALSE. Should roxygen be called to produce
documentation pages before package build?}
}
\description{
Create packages containing compiled elements
}
\details{

}
